# src/WebUI/Server/wwwroot/api/v1/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "CleanArchitecture API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Auctions/{id}": {
      "get": {
        "tags": [
          "Auctions"
        ],
        "operationId": "Auctions_GetAuction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionItemDetailsVM"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Auctions"
        ],
        "operationId": "Auctions_PutAuctionItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAuctionItemRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Auctions"
        ],
        "operationId": "Auctions_DeleteAuctionItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auctions/new": {
      "post": {
        "tags": [
          "Auctions"
        ],
        "operationId": "Auctions_AddAuction",
        "requestBody": {
          "x-name": "newAuctionItem",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuctionItemDTO"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuctionItemDTO": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "title",
          "description",
          "startDate",
          "endDate",
          "category",
          "startingBidAmount",
          "buyItNowPrice",
          "reservePrice",
          "deliveryMethod"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "publicId": {
            "type": "string",
            "format": "guid"
          },
          "title": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "category": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1
          },
          "startingBidAmount": {
            "type": "number",
            "format": "decimal"
          },
          "currentBidAmount": {
            "type": "number",
            "format": "decimal"
          },
          "buyItNowPrice": {
            "type": "number",
            "format": "decimal"
          },
          "reservePrice": {
            "type": "number",
            "format": "decimal"
          },
          "deliveryMethod": {
            "$ref": "#/components/schemas/DeliveryMethod"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "biddingHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BidDTO"
            }
          }
        }
      },
      "DeliveryMethod": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "None",
          "DeliveryByCourier",
          "PersonalPickUp"
        ],
        "enum": [
          0,
          1,
          2
        ]
      },
      "BidDTO": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "itemId": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number",
            "format": "decimal"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AuctionItemDetailsVM": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "auction": {
            "$ref": "#/components/schemas/AuctionItemDTO"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UpdateAuctionItemRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "category": {
            "type": "string"
          },
          "startingBidAmount": {
            "type": "number",
            "format": "decimal"
          },
          "currentBidAmount": {
            "type": "number",
            "format": "decimal"
          },
          "buyItNowPrice": {
            "type": "number",
            "format": "decimal"
          },
          "reservePrice": {
            "type": "number",
            "format": "decimal"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "deliveryMethod": {
            "$ref": "#/components/schemas/DeliveryMethod"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "biddingHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BidDTO"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Type into the textbox: Bearer {your JWT token}.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}